refactor: update import paths and remove deprecated files; add new services for AI assistant functionality
