+++ server/routes/summaries.py
@@ -6,7 +6,7 @@ import schemas.summary_schema as schemas
 import dependencies
 import services.summary_service as crud
 from file_reader_functions import extract_text
-from ai_assistant_summarization import summarize_text
+from services.ai_assistant_summarization import summarize_text
 
 router = APIRouter()
